web3 refactor (in progress)
