+++ lib/providermanager.js
@@ -23,9 +23,7 @@
  * @date 2014
  */
 
-var web3 = require('./web3'); 
 var jsonrpc = require('./jsonrpc');
-
 
 /**
  * Provider manager object prototype
